CI: los dos estrechamientos que quedaban rompiendo Windows
Segunda tanda de C4244, y esta vez los busque TODOS en local con clang
-Wconversion en vez de esperar otro viaje de dos minutos a la CI. Debi hacerlo
en el commit anterior.

1. ai_stub.hpp: `deficit` era int32_t, pero `player_stock` es int64_t, asi que
   la resta ya era de 64 bits y meterla en 32 era un truncado de verdad, no un
   quisquilloseo del compilador. Ahora es int64_t, que es el ancho honesto de
   la operacion.

2. bench_partida_larga.cpp: el indice de jugador es uint32_t y la mascara toma
   uint8_t. Cast explicito; kN_PLAYERS es 2.

Ninguno cambia comportamiento en gcc ni clang: los valores siempre cupieron.
Lo que cambiaba era que MSVC con -Werror no compilaba NADA.

Suite 51/51, cero avisos en local.

+++ addons/chunsa_sim/core/include/chunsa/ai_stub.hpp
@@ -771,7 +771,12 @@ inline AiTradeV1 ai_find_trade(const GameState& g, uint8_t player) noexcept {
                 missing = rr;
             }
             if (!clean || missing == 0xFFu) continue;
-            const int32_t deficit = rec.input[missing] - g.player_stock[player][missing];
+            // int64_t y no int32_t: `player_stock` ES int64_t, asi que la
+            // resta ya lo es, y meterla en un int32_t es un truncado que MSVC
+            // avisa (C4244) con razon. El tipo honesto es el ancho de la
+            // operacion; el filtro de la linea siguiente acota el valor util a
+            // (0, MARKET_LOT] de todos modos.
+            const int64_t deficit = rec.input[missing] - g.player_stock[player][missing];
             if (deficit <= 0 || deficit > MARKET_LOT) continue;
             int32_t precio = g.market_price_bp[player][missing];
             if (precio == 0) precio = MARKET_BASE_BP;           // sin inicializar = base
